Moved logo component to the correct folder, fixed linting errors, fixed css for firefox
