+++ client/.eslintrc.json
@@ -32,6 +32,7 @@
     "import/newline-after-import": "warn",
     "import/no-duplicates": "warn",
     "spaced-comment": ["warn", "always"],
-    "no-console": "warn"
+    "no-console": "warn",
+    "indent": 2
   }
 }
