+++ client/src/components/ThemeDropdown.tsx
@@ -2,57 +2,58 @@ import { useEffect, useState } from 'react';
 import Button from 'react-bootstrap/esm/Button';
 import Modal from 'react-bootstrap/Modal';
 import { AiFillSetting } from 'react-icons/ai';
+
 import { getContrastTextColor } from '../utils/colorBrightness'
 
 export const DEFAULT_THEME = {
-  /**
+    /**
    * The header component background color
    */
-  header: '#DFB479',
-  /**
+    header: '#DFB479',
+    /**
    * The header component text color
    */
-  headerText: '#212529',
-  /**
+    headerText: '#212529',
+    /**
    * The tutorial component background color
    */
-  tut: '#AED581',
-  /**
+    tut: '#AED581',
+    /**
    * The tutorial component text color
    */
-  tutText: '#212529',
-  /**
+    tutText: '#212529',
+    /**
    * The laboratory component background color
    */
-  lab: '#4FC3F7',
-  /**
+    lab: '#4FC3F7',
+    /**
    * The laboratory component text color
    */
-  labText: '#212529',
-  /**
+    labText: '#212529',
+    /**
    * The lecture component background color
    */
-  lec: '#FFB74D',
-  /**
+    lec: '#FFB74D',
+    /**
    * The lecture component text color
    */
-  lecText: '#212529',
-  /**
+    lecText: '#212529',
+    /**
    * The note component background color
    */
-  note: '#FF0000',
-  /**
+    note: '#FF0000',
+    /**
    * The note component text color
    */
-  noteText: '#FFFFFF',
-  /**
+    noteText: '#FFFFFF',
+    /**
    * The light text color
    */
-  lightText: '#FFFFFF',
-  /**
+    lightText: '#FFFFFF',
+    /**
    * The dark text color
    */
-  darkText: '#212529',
+    darkText: '#212529',
 };
 
 /**
@@ -73,28 +74,28 @@ type ColorInputProps = {
  * @returns The header color input component
  */
 const HeaderColorInput = ({ label, value, setValue }: ColorInputProps) => {
-  const bg = label.toLowerCase().includes(`header`) ? `bg-second` : `bg-${label.toLowerCase()}`;
-  return (
-    <div
-    className={`w-80 rounded ${bg}`}
-    style={{
-      display:'stretch',
-      height: 60,
-      padding: 10,
-      margin: 10,
-      alignItems: 'stretch',
-    }}>
-      <h6 className="m-0 text-bold">{label}</h6>
-      <input
-        type={'color'}
-        value={value}
-        onChange={(e) => {
-          setValue(e.target.value);
-        }}
-        style={{width:`100%`}}
-      />
-    </div>
-  );
+    const bg = label.toLowerCase().includes(`header`) ? `bg-second` : `bg-${label.toLowerCase()}`;
+    return (
+        <div
+            className={`w-80 rounded ${bg}`}
+            style={{
+                display:'stretch',
+                height: 60,
+                padding: 10,
+                margin: 10,
+                alignItems: 'stretch',
+            }}>
+            <h6 className="m-0 text-bold">{label}</h6>
+            <input
+                type={'color'}
+                value={value}
+                onChange={(e) => {
+                    setValue(e.target.value);
+                }}
+                style={{width:`100%`}}
+            />
+        </div>
+    );
 };
 
 /**
@@ -103,43 +104,43 @@ const HeaderColorInput = ({ label, value, setValue }: ColorInputProps) => {
  * @returns The color input component
  */
 const ColorInput = ({ label, value, setValue, noteLabel, noteValue, noteSetValue }: ColorInputProps) => {
-  const bg = label.toLowerCase().includes(`header`) ? `bg-second` : `bg-${label.toLowerCase()}`;
-  return (
-    <div
-    className={`lesson rounded ${bg}`}
-    style={{
-      width: '29%',
-      display:'inline-block',
-      padding: 10,
-      margin: 10,
-      justifyContent: 'space-between',
-      alignItems: 'stretch',
+    const bg = label.toLowerCase().includes(`header`) ? `bg-second` : `bg-${label.toLowerCase()}`;
+    return (
+        <div
+            className={`lesson rounded ${bg}`}
+            style={{
+                width: '29%',
+                display:'inline-block',
+                padding: 10,
+                margin: 10,
+                justifyContent: 'space-between',
+                alignItems: 'stretch',
       
-    }}>
-      <h6 className="text-bold">{label}</h6>
-      <input
-        type={'color'}
-        value={value}
-        onChange={(e) => {
-          setValue(e.target.value);
-        }}
-        style={{width:`100%`}}
-      />
-      {noteLabel && noteValue && noteSetValue && (
-        <h6 className="note bg-note p-1 text-bold rounded">
-          {noteLabel}
-          <input
-        type={'color'}
-        value={noteValue}
-        onChange={(e) => {
-          noteSetValue(e.target.value);
-        }}
-        style={{width:`100%`}}
-      />
-        </h6>
-      )}
-    </div>
-  );
+            }}>
+            <h6 className="text-bold">{label}</h6>
+            <input
+                type={'color'}
+                value={value}
+                onChange={(e) => {
+                    setValue(e.target.value);
+                }}
+                style={{width:`100%`}}
+            />
+            {noteLabel && noteValue && noteSetValue && (
+                <h6 className="note bg-note p-1 text-bold rounded">
+                    {noteLabel}
+                    <input
+                        type={'color'}
+                        value={noteValue}
+                        onChange={(e) => {
+                            noteSetValue(e.target.value);
+                        }}
+                        style={{width:`100%`}}
+                    />
+                </h6>
+            )}
+        </div>
+    );
 };
 
 /**
@@ -147,96 +148,96 @@ const ColorInput = ({ label, value, setValue, noteLabel, noteValue, noteSetValue
  * @returns The theme dropdown
  */
 export function ThemeDropdown() {
-  const [theme, setTheme] = useState(
-    localStorage.getItem('theme')
-      ? JSON.parse(localStorage.getItem('theme') as string)
-      : DEFAULT_THEME,
-  );
+    const [theme, setTheme] = useState(
+        localStorage.getItem('theme')
+            ? JSON.parse(localStorage.getItem('theme') as string)
+            : DEFAULT_THEME,
+    );
 
-  const [show, setShow] = useState(false);
+    const [show, setShow] = useState(false);
 
-  const handleClose = () => setShow(false);
-  const handleShow = () => setShow(true);
+    const handleClose = () => setShow(false);
+    const handleShow = () => setShow(true);
 
-  useEffect(() => {
-    document.documentElement.style.setProperty('--bg-tut', theme.tut);
-    document.documentElement.style.setProperty('--bg-tutText', theme.tutText);
-    document.documentElement.style.setProperty('--bg-lab', theme.lab);
-    document.documentElement.style.setProperty('--bg-labText', theme.labText);
-    document.documentElement.style.setProperty('--bg-lec', theme.lec);
-    document.documentElement.style.setProperty('--bg-lecText', theme.lecText);
-    document.documentElement.style.setProperty('--header', theme.header);
-    document.documentElement.style.setProperty('--bg-headerText', theme.headerText);
-    document.documentElement.style.setProperty('--bg-note', theme.note);
-    document.documentElement.style.setProperty('--bg-noteText', theme.noteText);
-    localStorage.setItem('theme', JSON.stringify(theme));
-  }, [theme]);
+    useEffect(() => {
+        document.documentElement.style.setProperty('--bg-tut', theme.tut);
+        document.documentElement.style.setProperty('--bg-tutText', theme.tutText);
+        document.documentElement.style.setProperty('--bg-lab', theme.lab);
+        document.documentElement.style.setProperty('--bg-labText', theme.labText);
+        document.documentElement.style.setProperty('--bg-lec', theme.lec);
+        document.documentElement.style.setProperty('--bg-lecText', theme.lecText);
+        document.documentElement.style.setProperty('--header', theme.header);
+        document.documentElement.style.setProperty('--bg-headerText', theme.headerText);
+        document.documentElement.style.setProperty('--bg-note', theme.note);
+        document.documentElement.style.setProperty('--bg-noteText', theme.noteText);
+        localStorage.setItem('theme', JSON.stringify(theme));
+    }, [theme]);
 
-  /**
+    /**
    * Sets the theme value
    * @param value The color value
    * @param key The theme name
    * @param keyText The according text theme name
    */
-  const setThemeValue = ( value: string, key: string, keyText?: string) => {
-    const themeCopy = { ...theme };
-    themeCopy[key] = value;
-    if (keyText !== undefined) {themeCopy[keyText] = getContrastTextColor(value);}
-    setTheme(themeCopy);
-  };
+    const setThemeValue = ( value: string, key: string, keyText?: string) => {
+        const themeCopy = { ...theme };
+        themeCopy[key] = value;
+        if (keyText !== undefined) {themeCopy[keyText] = getContrastTextColor(value);}
+        setTheme(themeCopy);
+    };
 
-  return (
-    <>
-      <Button
-        variant="primary"
-        onClick={handleShow}
-        className="text-bold bg-black">
-        <AiFillSetting />
-      </Button>
-      <Modal
-        className="modal-dark py-3 text-center text-light"
-        show={show}
-        onHide={handleClose}
-        style={{alignItems: 'center'}}>
-        <Modal.Header closeButton>
-          <Modal.Title>Theme settings</Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-        <div>
-              <HeaderColorInput
-              label={'Header'}
-              value={theme.header}
-              setValue={(e) => setThemeValue(e, 'header', 'headerText')}
-              />
-        </div>
-        <div>
-              <ColorInput
-              label={'Lec'}
-              value={theme.lec}
-              setValue={(e) => setThemeValue(e, 'lec', 'lecText')}/>
-              <ColorInput
-              label={'Lab'}
-              value={theme.lab}
-              setValue={(e) => setThemeValue(e, 'lab', 'labText')}/>
-              <ColorInput
-              label={'Tut'}
-              value={theme.tut}
-              setValue={(e) => setThemeValue(e, 'tut', 'tutText')}
-              noteLabel={'note'}
-              noteValue={theme.note}
-              noteSetValue={(e) => setThemeValue(e, 'note', 'noteText')}/>
-            </div>          
-        </Modal.Body>
-        <Modal.Footer>
-          <Button 
-            className={'w-100 bg-black'}
-            onClick={() => {
-              setTheme(DEFAULT_THEME);
-            }}>
+    return (
+        <>
+            <Button
+                variant="primary"
+                onClick={handleShow}
+                className="text-bold bg-black">
+                <AiFillSetting />
+            </Button>
+            <Modal
+                className="modal-dark py-3 text-center text-light"
+                show={show}
+                onHide={handleClose}
+                style={{alignItems: 'center'}}>
+                <Modal.Header closeButton>
+                    <Modal.Title>Theme settings</Modal.Title>
+                </Modal.Header>
+                <Modal.Body>
+                    <div>
+                        <HeaderColorInput
+                            label={'Header'}
+                            value={theme.header}
+                            setValue={(e) => setThemeValue(e, 'header', 'headerText')}
+                        />
+                    </div>
+                    <div>
+                        <ColorInput
+                            label={'Lec'}
+                            value={theme.lec}
+                            setValue={(e) => setThemeValue(e, 'lec', 'lecText')}/>
+                        <ColorInput
+                            label={'Lab'}
+                            value={theme.lab}
+                            setValue={(e) => setThemeValue(e, 'lab', 'labText')}/>
+                        <ColorInput
+                            label={'Tut'}
+                            value={theme.tut}
+                            setValue={(e) => setThemeValue(e, 'tut', 'tutText')}
+                            noteLabel={'note'}
+                            noteValue={theme.note}
+                            noteSetValue={(e) => setThemeValue(e, 'note', 'noteText')}/>
+                    </div>          
+                </Modal.Body>
+                <Modal.Footer>
+                    <Button 
+                        className={'w-100 bg-black'}
+                        onClick={() => {
+                            setTheme(DEFAULT_THEME);
+                        }}>
             Reset
-          </Button>
-        </Modal.Footer>
-      </Modal>
-    </>
-  );
+                    </Button>
+                </Modal.Footer>
+            </Modal>
+        </>
+    );
 }
