+++ client/src/hooks/useVictim.tsx
@@ -1,12 +1,13 @@
-import { VictimContext } from '../contexts/VictimContext';
 import { useContext } from 'react';
 
+import { VictimContext } from '../contexts/VictimContext';
+
 export default function useVictim() {
-  const context = useContext(VictimContext);
+    const context = useContext(VictimContext);
 
-  if (context === undefined) {
-    throw new Error('useVictim must be used within a <VictimProvider>!');
-  }
+    if (context === undefined) {
+        throw new Error('useVictim must be used within a <VictimProvider>!');
+    }
 
-  return context;
+    return context;
 }
