+++ client/src/utils/TimeUtil.ts
@@ -5,8 +5,8 @@ import { TimetableRecord } from '../services/DataService';
 const timeFormat = 'HH:mm';
 
 export function intervalFromLesson(lesson: TimetableRecord) {
-  return {
-    start: parse(lesson.startTime, timeFormat, 0),
-    end: parse(lesson.endTime, timeFormat, 0),
-  };
+    return {
+        start: parse(lesson.startTime, timeFormat, 0),
+        end: parse(lesson.endTime, timeFormat, 0),
+    };
 }
