+++ client/src/utils/intersections.ts
@@ -1,11 +1,11 @@
 import {
-  PersonData,
-  PersonInfo,
-  TimetableRecord,
+    PersonData,
+    PersonInfo,
+    TimetableRecord,
 } from '../services/DataService';
 
 const isInvalidNote = (note: string|undefined) => {
-  return !note || note === "" || (note !== "odd" && note !== "even");
+    return !note || note === "" || (note !== "odd" && note !== "even");
 }
 
 /**
@@ -20,13 +20,13 @@ export type LessonIntersections = {
  * Get an identification key for a given lesson.
  */
 export const getLessonKey = (lesson: TimetableRecord) => {
-  const key = `${lesson.day}-${lesson.title}-${lesson.type}-${lesson.room}-${lesson.startTime}-${lesson.endTime}`;
+    const key = `${lesson.day}-${lesson.title}-${lesson.type}-${lesson.room}-${lesson.startTime}-${lesson.endTime}`;
 
-  if (isInvalidNote(lesson.note)) {
-    return key;
-  }
+    if (isInvalidNote(lesson.note)) {
+        return key;
+    }
 
-  return `${key}-${lesson.note}`;
+    return `${key}-${lesson.note}`;
 };
 
 /**
@@ -35,25 +35,25 @@ export const getLessonKey = (lesson: TimetableRecord) => {
  * @param people
  */
 export const getLessonIntersectionsMap = (
-  people: PersonData[],
+    people: PersonData[],
 ): LessonIntersections => {
-  const intersections: LessonIntersections = {};
+    const intersections: LessonIntersections = {};
 
-  for (const person of people) {
-    for (const day of person.timetable) {
-      for (const lesson of day) {
-        const key = getLessonKey(lesson);
-        if (!intersections[key]) {
-          intersections[key] = [];
+    for (const person of people) {
+        for (const day of person.timetable) {
+            for (const lesson of day) {
+                const key = getLessonKey(lesson);
+                if (!intersections[key]) {
+                    intersections[key] = [];
+                }
+
+                intersections[key].push({
+                    id: person.id,
+                    name: person.name,
+                });
+            }
         }
-
-        intersections[key].push({
-          id: person.id,
-          name: person.name,
-        });
-      }
     }
-  }
 
-  return intersections;
+    return intersections;
 };
